commentlist

+++ ImPinker/Model/ArticleComment.cs
@@ -8,9 +8,9 @@ namespace ImModel
 {
     public class ArticleComment
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int ArticleId { get; set; }
+        public long ArticleId { get; set; }
 
         public int UserId { get; set; }
